Separate storage and aggregate on file level
To make it clear that every aggregate within a director has a
specialized Storage interface the definition and implementation of
Storage is separated into an own storage.go file.
