+++ sensor/memory/memory.go
@@ -33,7 +33,7 @@ func (gm getMemory) Get(g messages.Get) (interface{}, *messages.Failure, error) 
 	return &response, nil, nil
 }
 
-func New() handler.Holder{
+func New() handler.Holder {
 	return handler.FromGetter("memory", getMemory{
 		stats: mem.Get,
 	})
